Provide a new way to handle GET queries
A new "queryupdate" tag was created. This tag accepts at least an
argument of type QueryDict. Optionally it can have further arguments of
the form var=value. The optional arguments will be used to update the
QueryDict argument.

Sample usage:

{% queryupdate query page=2 foo=bar %}

The result will be:

?page=2&foo=bar

Assuming the original query is empty.

If bar is a context variable, then it will be resolved.