Doc string for the Paginator contructor
Fixed a small bug, when we change a page on a MessageListr. If we're
using the SORT extension we've already trimmed the flat_message_list to
contain only the needed message UID ou IMAP sequance numbers. Doing this
we avoid downloading the message information for all the messages on the
server. But, because we're changing pages, this formation is outdated so
we have to flag a new refresh of the messages. This is not optimal and
should be changed quickly.