Use EXISTS to get the mailbox message number
If the search program is "ALL", the EXISTS untagged response we get when
SELECTing the mailbox is number of messages in the mailbox. This way we
do not need to retrieve the message UIDs ou sequence numbers to
determine the number of messages.

This partially solves the problem identified in commit
ee479c28a4090d4af08269a2a8be340aec87f843 .