Get the References message header
We were only getting the References message header when listing the
messages, since this header is needed to do client side threading.
However this header is also needed to reply to messages (the replied
message References header must be equal to the answered message
References plus the replied Message-ID header of the old message).